// Backend/src/db/types/public/Paises.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.Paises */
export type PaisesCodPais = number & { __brand: 'public.Paises' };

/** Represents the table public.Paises */
export default interface Paises {
  codPais: PaisesCodPais;

  pais: string;

  sigla: string;

  ddi: string | null;

  moeda: string | null;

  dtCriacao: Date;

  dtEdicao: Date;

  codUser: number | null;

  isActive: boolean;
}

/** Represents the initializer for the table public.Paises */
export interface PaisesInitializer {
  codPais?: PaisesCodPais;

  pais: string;

  sigla: string;

  ddi?: string | null;

  moeda?: string | null;

  /** Default value: CURRENT_TIMESTAMP */
  dtCriacao?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  dtEdicao?: Date;

  codUser?: number | null;

  /** Default value: true */
  isActive?: boolean;
}

/** Represents the mutator for the table public.Paises */
export interface PaisesMutator {
  codPais?: PaisesCodPais;

  pais?: string;

  sigla?: string;

  ddi?: string | null;

  moeda?: string | null;

  dtCriacao?: Date;

  dtEdicao?: Date;

  codUser?: number | null;

  isActive?: boolean;
}